Start Calendar
I started implementation for the calendar system. I have done some research and decided to implement Full Calendar as well as google calendar. I have made it where I can add events on google calendar and it will appear on the webpage, but have not added a way to create events from the webpage. That is my next step.