Login
I initially started working on the calendar system. This was until I got to a point where I realised I cannot interact with the calendar unless I am a user. So I must implement the login and registering system.  I started in a diffrent codebase so that I could get a grasp on how to implement this. I was successful in implementing a registering and loggin in sytem there, but when I tried to convert it to match our existing code it did not work. I messed with this for a couple hours, but I think it may be do to the package.json. I will be sending my teamates a zipped folder with the working iteration.