First Testing FIle push
I fixed the issues that we were having in the original file by using some router express and requiring. I also fixed the issue with the current login system. I added the random verse functionality, and got it working, I also got the verse searcher implemented into this as well. I was able to successfully make a prayerwall post and display it. I need to do work on the prayer wall tommorow, but a post and displaying is avaliable now.